refactor: migrate no-deprecated rule
